projects/Barky/requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pep517==0.10.0
pip-tools==6.0.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.15.0
regex==2021.3.17
requests==2.25.1
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
